+++ database/migrations/2025_05_25_011303_create_suppliers_table.php
@@ -11,29 +11,31 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('suppliers', function (Blueprint $table) {
-            $table->id();
+        if (!Schema::hasTable('suppliers')) {
+            Schema::create('suppliers', function (Blueprint $table) {
+                $table->id();
 
-            $table->foreignId('identity_id')
-                ->constrained()
-                ->onDelete('cascade');
+                $table->foreignId('identity_id')
+                    ->constrained()
+                    ->onDelete('cascade');
 
-            $table->string('document_number')
-                ->unique();
+                $table->string('document_number')
+                    ->unique();
 
-            $table->string('name');
+                $table->string('name');
 
-            $table->string('address')
-                ->nullable();
+                $table->string('address')
+                    ->nullable();
 
-            $table->string('email')
-                ->nullable();
+                $table->string('email')
+                    ->nullable();
 
-            $table->string('phone')
-                ->nullable();
+                $table->string('phone')
+                    ->nullable();
 
-            $table->timestamps();
-        });
+                $table->timestamps();
+            });
+        }
     }
 
     /**
